% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classifyFunk.r
\name{classify}
\alias{classify}
\title{A function for internal use that formats a generic function for internal use}
\usage{
classify(fun, numPars = 1)
}
\arguments{
\item{fun}{the function to internalize}

\item{numPars}{number of parameters used by fun}
}
\description{
A function for internal use that formats a generic function for internal use
}
