% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_modelFunk.r, R/10_ddModelFunk.r,
%   R/20_prodModelFunk.r
\docType{data}
\name{dLikes}
\alias{dLikes}
\title{A list of functions for internal use that formats a generic function for internal use}
\format{
An object of class \code{list} of length 2.

An object of class \code{list} of length 2.

An object of class \code{list} of length 2.
}
\usage{
dLikes

dLikes

dLikes
}
\arguments{
\item{self}{a ddModel object}

\item{data}{data to be used with likelihood}
}
\description{
A list of functions for internal use that formats a generic function for internal use

A list of functions for internal use that formats a generic function for internal use

A list of functions for internal use that formats a generic function for internal use
}
\keyword{datasets}
