% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classifyFunk.r
\name{dNdt_classify}
\alias{dNdt_classify}
\title{A function for internal use that formats the dNdt function for internal use}
\usage{
dNdt_classify(fun, numPars = 2)
}
\arguments{
\item{fun}{the function to internalize}

\item{numPars}{number of parameters used by fun}
}
\description{
A function for internal use that formats the dNdt function for internal use
}
