% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_ddOutputFunk.r, R/20_prodOutputFunk.r
\name{makeTransparent}
\alias{makeTransparent}
\title{A function primarily for internal use that make colors transparent.}
\usage{
makeTransparent(someColor, alpha = 100)

makeTransparent(someColor, alpha = 100)
}
\arguments{
\item{someColor}{a color given however R handles a color}

\item{alpha}{the alpha channel value that wil be added}
}
\value{
a version of someColor with the alpha channel added

a version of someColor with the alpha channel added
}
\description{
A function primarily for internal use that make colors transparent.

A function primarily for internal use that make colors transparent.
}
