% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classifyFunk.r
\name{N0_classify}
\alias{N0_classify}
\title{A function for internal use that formats the N0 function for internal use}
\usage{
N0_classify(fun)
}
\arguments{
\item{fun}{the function to internalize}
}
\description{
A function for internal use that formats the N0 function for internal use
}
