% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_ddOptimizeFunk.r, R/20_prodOptimizeFunk.r
\name{optimize}
\alias{optimize}
\title{A function for internal usage within the ddModel class to optimize model parameter given the data provided}
\usage{
optimize(
  data,
  parNames,
  lower,
  upper,
  method = self$OPT_method,
  cov = F,
  fitQ = T,
  gaBoost = F,
  persistFor = 50,
  control = list()
)

optimize(
  data,
  parNames,
  lower,
  upper,
  method = self$OPT_method,
  cov = F,
  fitQ = T,
  gaBoost = F,
  persistFor = 50,
  control = list()
)
}
\arguments{
\item{data}{A vector of data used to fit specified model.}

\item{parNames}{A vector of strings matching the names of the parameters to be optimized}

\item{lower}{A vector of lower bounds for searching parameter space.}

\item{upper}{A vector of upper bounds for searching parameter space.}

\item{method}{A string optionally defining the method of optimization to be handed to optim (default 'L-BFGS-B')}

\item{cov}{A logical optionally indicating if hessian should be computed and inverted in optimization process}

\item{fitQ}{A logical indicating whether to find the MLE of log(q):lq via profile MLE, or if False don't change the initialized value of lq.}

\item{gaBoost}{A logical optionally (default F) indicating if a persistent genetic algorithm should be used to assist local optimization. Genetic algorithm repeates until first and second finite difference derivatives are successful and hessian is inverted. Optionally gaBoost may be given as a list containting names values for list(popSize, maxiter, run).}

\item{persistFor}{A numeric indicating how many iterations of optimization tryCatch to engange in.}

\item{control}{Additional control parameters to be passed to optim}
}
\value{
Optimization objects are returned. Parameters values are updated inplace. rsCov is defined to self.

Optimization objects are returned. Parameters values are updated inplace. rsCov is defined to self.
}
\description{
A function for internal usage within the ddModel class to optimize model parameter given the data provided

A function to optimize model parameter given the data provided
}
