% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_parSelfParFunk.r
\name{parToSelf}
\alias{parToSelf}
\title{A function primarily for internal use}
\usage{
parToSelf(parValues)
}
\arguments{
\item{parValues}{values of parameters given as named objects to be registered with self}
}
\description{
A function primarily for internal use
}
