% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_ddOutputFunk.r, R/20_prodOutputFunk.r
\name{plotBand}
\alias{plotBand}
\title{A function for internal use to plot uncertainty bands}
\usage{
plotBand(prob = 0.95, col = "black", alpha = 100)

plotBand(prob = 0.95, col = "black", alpha = 100)
}
\arguments{
\item{prob}{Size of the uncertainty bands as defined as posterior probability}

\item{col}{col}

\item{alpha}{alpha}
}
\description{
A function for internal use to plot uncertainty bands

A function to plot uncertainty bands
}
