% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_ddOutputFunk.r, R/20_prodOutputFunk.r
\name{plotMean}
\alias{plotMean}
\title{A function for internal use to plot the mean of the dynamics}
\usage{
plotMean(col = "black", alpha = 100, lwd = 3, add = F)

plotMean(col = "black", alpha = 100, lwd = 3, add = F)
}
\arguments{
\item{col}{}

\item{alpha}{}

\item{lwd}{}

\item{add}{}
}
\description{
A function for internal use to plot the mean of the dynamics

A function to plot the mean of the dynamics
}
