% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_ddOutputFunk.r
\name{plotQuan}
\alias{plotQuan}
\title{A function for internal use to plot an abitratry transformation of model paratmeters thru time}
\usage{
plotQuan(
  quan = function(B) {
     B
 },
  col = "black",
  alpha = 100,
  lwd = 3,
  add = F,
  ...
)
}
\arguments{
\item{quan}{a quantity to plot given as a function}

\item{col}{}

\item{alpha}{}

\item{lwd}{}

\item{add}{}
}
\description{
A function for internal use to plot an abitratry transformation of model paratmeters thru time
}
