% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_ddOutputFunk.r, R/20_prodOutputFunk.r
\name{plotRS}
\alias{plotRS}
\title{A function for internal use to plot repeated sampling posterior-like parameter distributions}
\usage{
plotRS(m = 10^4, sample = F, save = F)

plotRS(m = 10^4, sample = F, save = F)
}
\arguments{
\item{m}{the number of samples from the repeated sampling distribution}

\item{sample}{a boolean to indicate if samples should be returned}

\item{save}{a boolean to indicate if samples should be saved}
}
\description{
A function for internal use to plot repeated sampling posterior-like parameter distributions

A function to plot repeated sampling posterior-like parameter distributions
}
