% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_ddOutputFunk.r, R/20_prodOutputFunk.r
\name{printSelf}
\alias{printSelf}
\title{A function for internal use to print}
\usage{
printSelf(ins, outs = c())

printSelf(ins, outs = c())
}
\arguments{
\item{ins}{ins}

\item{outs}{outs}
}
\description{
A function for internal use to print

A function to print
}
