% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22_prodClass0.1.1.r
\name{prodModel}
\alias{prodModel}
\title{The main production modeling class definition}
\value{
an initializer for the prodModel class
}
\description{
The main production modeling class definition

The main production modeling class definition
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{N}}{a vector of population numbers/biomass}

\item{\code{N0}}{a scalar initial condition for numbers/biomass}

\item{\code{N0Funk}}{a function defining the equilibrium vigin numbers as a function of the productivity parameters}

\item{\code{time}}{a vector of times for which to integrate the dynamics to}

\item{\code{lsdo}}{a scalar of the log scale residual sd}

\item{\code{lq}}{a scalar of the log proportionality constant between cpue and N}

\item{\code{model}}{a list of strings indicating the residual model, etc.}

\item{\code{ODE_method}}{a string specifing the integration method to be used by dede}

\item{\code{OPT_method}}{a string specifing the optimization method to be used by optim}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProdModel-optimize}{\code{prodModel$optimize()}}
\item \href{#method-ProdModel-printer}{\code{prodModel$printer()}}
\item \href{#method-ProdModel-plotMean}{\code{prodModel$plotMean()}}
\item \href{#method-ProdModel-plotBand}{\code{prodModel$plotBand()}}
\item \href{#method-ProdModel-plotRS}{\code{prodModel$plotRS()}}
\item \href{#method-ProdModel-new}{\code{prodModel$new()}}
\item \href{#method-ProdModel-iterate}{\code{prodModel$iterate()}}
\item \href{#method-ProdModel-printSelf}{\code{prodModel$printSelf()}}
\item \href{#method-ProdModel-save}{\code{prodModel$save()}}
\item \href{#method-ProdModel-load}{\code{prodModel$load()}}
\item \href{#method-ProdModel-like}{\code{prodModel$like()}}
\item \href{#method-ProdModel-clone}{\code{prodModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProdModel-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-ProdModel-optimize}{}}}
\subsection{Method \code{optimize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prodModel$optimize(
  data,
  parNames,
  lower,
  upper,
  method = self$OPT_method,
  cov = F,
  fitQ = T,
  gaBoost = F,
  persistFor = 50,
  control = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProdModel-printer"></a>}}
\if{latex}{\out{\hypertarget{method-ProdModel-printer}{}}}
\subsection{Method \code{printer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prodModel$printer(ins, outs = c())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProdModel-plotMean"></a>}}
\if{latex}{\out{\hypertarget{method-ProdModel-plotMean}{}}}
\subsection{Method \code{plotMean()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prodModel$plotMean(col = "black", alpha = 100, lwd = 3, add = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProdModel-plotBand"></a>}}
\if{latex}{\out{\hypertarget{method-ProdModel-plotBand}{}}}
\subsection{Method \code{plotBand()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prodModel$plotBand(prob = 0.95, col = "black", alpha = 100)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProdModel-plotRS"></a>}}
\if{latex}{\out{\hypertarget{method-ProdModel-plotRS}{}}}
\subsection{Method \code{plotRS()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prodModel$plotRS(m = 10^4, sample = F, save = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProdModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProdModel-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of the ddModel class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prodModel$new(N0 = NA, time = NA, dNdt = NA, N0Funk = NA, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{N0}}{a scalar initial condition for numbers}

\item{\code{time}}{a vector of times for which to integrate the dynamics to}

\item{\code{dNdt}}{a function that evaluates the delay differential equation that defines numbers/biomass dynamics.}

\item{\code{N0Funk}}{a function defining the equilibrium vigin numbers as a function of the productivity parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of the prodModel class, for which data can be provided, parameters optimized and quantities plotted.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProdModel-iterate"></a>}}
\if{latex}{\out{\hypertarget{method-ProdModel-iterate}{}}}
\subsection{Method \code{iterate()}}{
A function to integrate the dynamics equations over the times in 'time'
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prodModel$iterate(method = self$ODE_method)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{an optional string specifing the integration method to be used by dede}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL B and N are updated interally.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProdModel-printSelf"></a>}}
\if{latex}{\out{\hypertarget{method-ProdModel-printSelf}{}}}
\subsection{Method \code{printSelf()}}{
A function to optimize model parameter given the data provided


A function to print


The main function for printing this class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prodModel$printSelf(ins = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ins}}{ins}

\item{\code{ins}}{ins}

\item{\code{data}}{A vector of data used to fit specified model.}

\item{\code{parNames}}{A vector of strings matching the names of the parameters to be optimized}

\item{\code{lower}}{A vector of lower bounds for searching parameter space.}

\item{\code{upper}}{A vector of upper bounds for searching parameter space.}

\item{\code{method}}{A string optionally defining the method of optimization to be handed to optim (default 'L-BFGS-B')}

\item{\code{cov}}{A logical optionally indicating if hessian should be computed and inverted in optimization process}

\item{\code{fitQ}}{A logical indicating whether to find the MLE of log(q):lq via profile MLE, or if False don't change the initialized value of lq.}

\item{\code{gaBoost}}{A logical optionally (default F) indicating if a persistent genetic algorithm should be used to assist local optimization. Genetic algorithm repeates until first and second finite difference derivatives are successful and hessian is inverted. Optionally gaBoost may be given as a list containting names values for list(popSize, maxiter, run).}

\item{\code{persistFor}}{A numeric indicating how many iterations of optimization tryCatch to engange in.}

\item{\code{control}}{Additional control parameters to be passed to optim}

\item{\code{outs}}{outs}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Optimization objects are returned. Parameters values are updated inplace. rsCov is defined to self.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProdModel-save"></a>}}
\if{latex}{\out{\hypertarget{method-ProdModel-save}{}}}
\subsection{Method \code{save()}}{
A function to plot the mean of the dynamics


A function to plot uncertainty bands


A function to plot repeated sampling posterior-like parameter distributions


A function to save the ddModel class as an rds file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prodModel$save(fileName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fileName}}{A string defining the name and path of the file to be saved. It should probably end with the .rds extension.}

\item{\code{col}}{A color given however R colors are given}

\item{\code{alpha}}{An alpha level between 0 and 255}

\item{\code{lwd}}{Line width}

\item{\code{add}}{A boolean indicating if the plot should be added to the existing device or not}

\item{\code{prob}}{Size of the uncertainty bands as defined as posterior probability}

\item{\code{col}}{A color given however R colors are given}

\item{\code{alpha}}{An alpha level between 0 and 255}

\item{\code{m}}{The number of samples from the repeated sampling distribution}

\item{\code{sample}}{A boolean to indicate if samples should be returned}

\item{\code{save}}{A boolean to indicate if samples should be saved}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProdModel-load"></a>}}
\if{latex}{\out{\hypertarget{method-ProdModel-load}{}}}
\subsection{Method \code{load()}}{
A function to read in a ddModel class from .rds file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prodModel$load(fileName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fileName}}{A string of the name and path of the file to be loaded.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProdModel-like"></a>}}
\if{latex}{\out{\hypertarget{method-ProdModel-like}{}}}
\subsection{Method \code{like()}}{
A function to calculate the likelihood of the model given the data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prodModel$like(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A vector of the given index data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProdModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProdModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prodModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
