% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_parSelfParFunk.r
\name{selfToPar}
\alias{selfToPar}
\title{A function primarily for internal use}
\usage{
selfToPar(parNames)
}
\arguments{
\item{selfToPar}{}
}
\value{
parameter values taken from self
}
\description{
A function primarily for internal use
}
